requirements.txt
gradio
Pillow
numpy
trimesh